downloader/script/requirements.txt
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
exceptiongroup==1.2.2
git-filter-repo==2.47.0
h11==0.14.0
humanfriendly==10.0
idna==3.7
outcome==1.3.0.post0
packaging==24.2
pycparser==2.22
pyreadline3==3.5.4
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.0
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
wsproto==1.2.0
chardet==5.2.0
yt-dlp>=2025.03.21
python-ffmpeg==2.0.12